Added method prototypes for selecting stories and adjusting game state
Currently stories can be selected but not stored in the WeeklyBroadcast
array. There is a submit button which check that the broadcast is full
and then runs the effect of all stories. There is currently no way to
store weekly stories, and there are currently no checks to find the
negative effects of not running a story.